Consolidate Commerce types into types file
These commerce types were being defined and exported sporadically
throughout various lib, machine, and component files. I thought it would
be a lot cleaner to promote all of them to the `types.ts` file and then
import them as needed from there.
