+++ src/components/pages/landing/footer/index.tsx
@@ -6,13 +6,12 @@ import axios from 'axios'
 import * as Yup from 'yup'
 import {FormikProps, useFormik} from 'formik'
 import {useCommerceMachine} from 'hooks/use-commerce-machine'
-import {Coupon} from 'machines/commerce-machine'
 import {track} from 'utils/analytics'
 import emailIsValid from 'utils/email-is-valid'
 import {useViewer} from 'context/viewer-context'
 import ParityCouponMessage from 'components/pricing/parity-coupon-message'
 import {PlanPrice} from 'components/pricing/select-plan-new/index'
-import {StripeAccount} from 'types'
+import {Coupon, StripeAccount} from 'types'
 import stripeCheckoutRedirect from 'api/stripe/stripe-checkout-redirect'
 import Countdown from 'components/pricing/countdown'
 import {fromUnixTime} from 'date-fns'
