+++ src/components/pages/lessons/overlay/go-pro-cta-overlay.tsx
@@ -1,12 +1,11 @@
 import React, {FunctionComponent} from 'react'
 import Link from 'next/link'
 import ProMemberFeatures from 'components/pro-member-features'
-import {LessonResource} from 'types'
+import {Coupon, LessonResource} from 'types'
 import {track} from 'utils/analytics'
 import {useCommerceMachine} from 'hooks/use-commerce-machine'
 import {PlanPrice} from 'components/pricing/select-plan-new/index'
 import SmallParityCouponMessage from './small-parity-coupon-message'
-import {Coupon} from 'machines/commerce-machine'
 import get from 'lodash/get'
 import isEmpty from 'lodash/isEmpty'
 import first from 'lodash/first'
